Move test to a new file
